OPSMN-3746:System lockdown to disable any activity on the OpenSpecimen instance

+++ WEB-INF/src/com/krishagni/catissueplus/core/auth/services/impl/UserAuthenticationServiceImpl.java
@@ -31,6 +31,7 @@ import com.krishagni.catissueplus.core.common.events.RequestEvent;
 import com.krishagni.catissueplus.core.common.events.ResponseEvent;
 import com.krishagni.catissueplus.core.common.events.UserSummary;
 import com.krishagni.catissueplus.core.common.util.AuthUtil;
+import com.krishagni.catissueplus.core.common.util.ConfigUtil;
 import com.krishagni.catissueplus.core.common.util.Status;
 
 public class UserAuthenticationServiceImpl implements UserAuthenticationService {
@@ -54,6 +55,10 @@ public class UserAuthenticationServiceImpl implements UserAuthenticationService 
 		try {
 			user = daoFactory.getUserDao().getUser(loginDetail.getLoginName(), loginDetail.getDomainName());
 			
+			if (!user.isAdmin() && ConfigUtil.getInstance().getBoolSetting("administrative", "system_lockdown", false)) {
+				throw OpenSpecimenException.userError(AuthErrorCode.SYSTEM_LOCKDOWN);
+			}
+
 			if (user == null) {
 				throw OpenSpecimenException.userError(AuthErrorCode.INVALID_CREDENTIALS);
 			}
@@ -103,10 +108,14 @@ public class UserAuthenticationServiceImpl implements UserAuthenticationService 
 			if (authToken == null) {
 				throw OpenSpecimenException.userError(AuthErrorCode.INVALID_TOKEN);
 			}
-			
+
 			User user = authToken.getUser();
 			long timeSinceLastApiCall = auditService.getTimeSinceLastApiCall(user.getId(), token);
 			int tokenInactiveInterval = AuthConfig.getInstance().getTokenInactiveIntervalInMinutes();
+			if (!user.isAdmin() && ConfigUtil.getInstance().getBoolSetting("administrative", "system_lockdown", false)) {
+				throw OpenSpecimenException.userError(AuthErrorCode.SYSTEM_LOCKDOWN);
+			}
+
 			if (timeSinceLastApiCall > tokenInactiveInterval) {
 				daoFactory.getAuthDao().deleteAuthToken(authToken);
 				throw OpenSpecimenException.userError(AuthErrorCode.TOKEN_EXPIRED);
